Authorization class added with static method. Flash messages and edit/delete forms refactored

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -22,7 +24,7 @@ class ListingController
      */
     public function index()
     {
-        $listings = $this->db->query("SELECT * FROM listings")->fetchAll();
+        $listings = $this->db->query("SELECT * FROM listings ORDER BY created_at DESC")->fetchAll();
 
         loadView("listings/index", [
             "listings" => $listings,
@@ -95,7 +97,7 @@ class ListingController
             array_flip($allowedFields)
         );
 
-        $newListingData["user_id"] = 1;
+        $newListingData["user_id"] = Session::get("user")["id"];
 
         $newListingData = array_map("sanitize", $newListingData);
 
@@ -112,14 +114,14 @@ class ListingController
 
         foreach ($requireFields as $field) {
             if (
-                empty($newListingData[$field]) ||
+                empty ($newListingData[$field]) ||
                 !Validation::string($newListingData[$field])
             ) {
                 $errors[$field] = ucfirst($field) . " is required";
             }
         }
 
-        if (!empty($errors)) {
+        if (!empty ($errors)) {
             // Reload view with errors
             loadView("listings/create", [
                 "errors" => $errors,
@@ -151,6 +153,8 @@ class ListingController
 
             $this->db->query($query, $newListingData);
 
+            Session::setFlashMessage("success_message", "Listing created successfully");
+
             redirect("/listings");
         }
     }
@@ -173,21 +177,28 @@ class ListingController
             ->query("SELECT * FROM listings WHERE id = :id", $params)
             ->fetch();
 
+        // Check if listings exists
         if (!$listing) {
             ErrorController::notFound("Listing not found");
             return;
         }
 
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage("error_message", "You are not authorized to delete this listing");
+            return redirect("/listings/" . $listing->id);
+        }
+
         $this->db->query("DELETE FROM listings WHERE id = :id", $params);
 
         // Set flash message
-        $_SESSION["success_message"] = "Listing deleted successfully";
+        Session::setFlashMessage("success_message", "Listing deleted successfully");
 
         redirect("/listings");
     }
 
     /**
-     * Show the lsiting edit form
+     * Show the listing edit form
      *
      * @param array @params
      * @return void
@@ -208,6 +219,12 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound("Listing not found");
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage("error_message", "You are not authorized to update this listing");
+            return redirect("/listings/" . $listing->id);
         }
 
         loadView("listings/edit", [
@@ -237,6 +254,12 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound("Listing not found");
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage("error_message", "You are not authorized to update this listing");
+            return redirect("/listings/" . $listing->id);
         }
 
         $allowedFields = [
@@ -273,14 +296,14 @@ class ListingController
 
         foreach ($requiredFields as $field) {
             if (
-                empty($updateValues[$field]) ||
+                empty ($updateValues[$field]) ||
                 !Validation::string($updateValues[$field])
             ) {
                 $errors[$field] = ucfirst($field) . " is required";
             }
         }
 
-        if (!empty($errors)) {
+        if (!empty ($errors)) {
             loadView("listings/edit", [
                 "listing" => $listing,
                 "errors" => $errors,
@@ -301,7 +324,8 @@ class ListingController
             $updateValues["id"] = $id;
             $this->db->query($updateQuery, $updateValues);
 
-            $_SESSION["success_message"] = "Listing Updated";
+            // Set flash message
+            Session::setFlashMessage("success_message", "Listing updated successfully");
 
             redirect("/listings/" . $id);
         }
